🔨 separate state from rendering for horizontal color legends
