+++ packages/@ourworldindata/grapher/src/horizontalColorLegend/HorizontalColorLegendConstants.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_NUMERIC_BIN_SIZE = 10
+export const DEFAULT_NUMERIC_BIN_STROKE = "#333"
+export const DEFAULT_NUMERIC_BIN_STROKE_WIDTH = 0.3
+export const DEFAULT_TEXT_COLOR = "#111"
+export const DEFAULT_TICK_SIZE = 3
+
+export const CATEGORICAL_BIN_MIN_WIDTH = 20
+export const SPACE_BETWEEN_CATEGORICAL_BINS = 7
+export const MINIMUM_LABEL_DISTANCE = 5
